requirements.txt
perlin-noise==1.12
Pillow==9.4.0
